Refactor: reducer 이름 변경 및 폴더 변경
