--- src/features/testSlice/testSlice.ts
@@ -1,1 +1,0 @@
-export {};
